requirements.txt
faiss-cpu
flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
langchain
langchain-community
langchain-google-genai
langchain-openai
openai
pypdf
python-dateutil
python-dotenv
pytz
requests
tiktoken
waitress
cryptography
pandas
gspread
oauth2client
